Implement Arena with limits to movement
This implements the new `Arena` class, which specifies limits to the
movement of knights: any knight moving off the edge of the arena is
drowned.

This also updates `Piece` initialisation to require an arena since
pieces must know the limits for movement. This tight coupling is perhaps
not ideal; should consider if limits can be enforced in a more
lightweight way to eliminate this.