Remove piece awareness of arena
Pieces are now unaware that they are in an arena, removing the
requirement for an arena for initialisation. Instead, positions are
_optionally_ checked to be within limits by explicitly passing limits
seperately rather than pulling from the arena.

This provides no functional change, it is merely a stylistic/design
decision to reduce tight coupling between objects (and eliminate the
circular import dependency between `game` and `pieces`)