Implement single instruction
The arena method `run_instruction` ties together the piece interactions
on knight movement.

This implements a partial movement validity check in order to appease
mypy, however it also supercedes that inside of knight and is untested.

Generally, this instruction is not well implemented because much of the
position checking logic which the knight currently carries out should
actually occur at the arena --- there is considerable scope for
refactoring here.